Suppress output from jsonlint
We're using jsonlint to make sure that the package file is valid
JSON so we don't need any of its output. What would be nice however
is to still let through the error output because it gives hints as
to what is wrong in the package file.

We would however need to do one of these:

1- use "chronic" (in Debian: apt-get install moreutils)
2- modify jsonlint to add a "validate only" cli option
3- modify jsonlint to output errors to stderr instead of stdout

Of these three options, #3 looks easiest and #2 seems like the
ideal one.